Collapse migrationTool and migrator directories, rename to interfaces directory
