+++ examples/utils/migration-tools/src/interfaces/migratableModel.ts
