+++ examples/utils/migration-tools/src/migrationTool.ts
@@ -27,7 +27,7 @@ import type {
 	IMigrationTool,
 	IMigrationToolEvents,
 	MigrationState,
-} from "../migrationInterfaces/index.js";
+} from "./interfaces/index.js";
 
 const consensusRegisterCollectionId = "consensus-register-collection";
 const pactMapId = "pact-map";
