--- examples/utils/migration-tools/src/migrationTool/index.ts
@@ -1,6 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
- * Licensed under the MIT License.
- */
-
-export { MigrationToolFactory } from "./migrationTool.js";
